move sync routes to a subfolder
