+++ worker/sync/src/api/sync/$sku/asset/upload_url.ts
@@ -1,5 +1,9 @@
 import { Cloudflare } from "cloudflare";
-import { ErrorResponseSchema, ErrorResponses, AppArgs } from "../../../router";
+import {
+  ErrorResponseSchema,
+  ErrorResponses,
+  AppArgs,
+} from "../../../../router";
 import { createRoute, RouteHandler } from "@hono/zod-openapi";
 import { z } from "zod/v4";
 import {
@@ -7,9 +11,9 @@ import {
   verifySignature,
   VerifySignatureHeadersSchema,
   verifyUser,
-} from "../../../utils/verify";
+} from "../../../../utils/verify";
 import { ImageAssetMeta } from "@referee-fyi/share";
-import { getAssetMeta, setAssetMeta } from "../../../utils/data";
+import { getAssetMeta, setAssetMeta } from "../../../../utils/data";
 export const ParamsSchema = z.object({
   sku: z.string(),
 });
