+++ worker/sync/src/api/sync/$sku/incident/patch.ts
@@ -1,4 +1,8 @@
-import { ErrorResponseSchema, ErrorResponses, AppArgs } from "../../../router";
+import {
+  ErrorResponseSchema,
+  ErrorResponses,
+  AppArgs,
+} from "../../../../router";
 import { createRoute, RouteHandler } from "@hono/zod-openapi";
 import { z } from "zod/v4";
 import {
@@ -6,44 +10,64 @@ import {
   verifySignature,
   VerifySignatureHeadersSchema,
   verifyUser,
-} from "../../../utils/verify";
-import { WebSocketSender } from "@referee-fyi/share";
-import { getUser } from "../../../utils/data";
-export const QuerySchema = z.object({
-  id: z.string(),
-});
-
+} from "../../../../utils/verify";
+import {
+  Incident,
+  INCIDENT_IGNORE,
+  IncidentSchema,
+  WebSocketSender,
+} from "@referee-fyi/share";
+import { getUser } from "../../../../utils/data";
+import { mergeLWW } from "@referee-fyi/consistency";
 export const ParamsSchema = z.object({
   sku: z.string(),
 });
 
+export const BodySchema = IncidentSchema;
+
 export const SuccessResponseSchema = z
   .object({
     success: z.literal(true),
-    data: z.object({}),
+    data: IncidentSchema,
   })
   .meta({
-    id: "DeleteIncidentResponse",
+    id: "PatchIncidentResponse",
   });
 
 export const route = createRoute({
-  method: "delete",
+  method: "patch",
   path: "/api/{sku}/incident",
   tags: ["Incident"],
-  summary: "Delete an incident.",
+  summary: "Edit an incident.",
   hide: process.env.WRANGLER_ENVIRONMENT === "production",
   middleware: [verifySignature, verifyUser, verifyInvitation],
   request: {
     headers: VerifySignatureHeadersSchema,
-    query: QuerySchema,
     params: ParamsSchema,
+    body: {
+      required: true,
+      description: "Incident body to edit",
+      content: {
+        "application/json": {
+          schema: BodySchema,
+        },
+      },
+    },
   },
   responses: {
     200: {
-      description: "Incident deleted successfully",
+      description: "Incident created successfully",
       content: {
         "application/json": {
           schema: SuccessResponseSchema,
+        },
+      },
+    },
+    409: {
+      description: "Incident edit conflict",
+      content: {
+        "application/json": {
+          schema: ErrorResponseSchema,
         },
       },
     },
@@ -53,7 +77,6 @@ export const route = createRoute({
 
 export type Route = typeof route;
 export const handler: RouteHandler<Route, AppArgs> = async (c) => {
-  const { id } = c.req.valid("query");
   const verifyInvitation = c.get("verifyInvitation");
   if (!verifyInvitation) {
     return c.json(
@@ -69,11 +92,28 @@ export const handler: RouteHandler<Route, AppArgs> = async (c) => {
     );
   }
 
+  const incident = c.req.valid("json") as Incident;
+
   const stub = c.env.INCIDENTS.get(
     c.env.INCIDENTS.idFromString(verifyInvitation.instance.secret)
   );
 
-  await stub.deleteIncident(id);
+  const deleted = await stub.getDeletedIncidents();
+  if (deleted.has(incident.id)) {
+    return c.json(
+      {
+        success: false,
+        code: "PatchIncidentDeleted",
+        error: {
+          name: "ValidationError",
+          message: "This incident has already been deleted.",
+        },
+      } as const satisfies z.infer<typeof ErrorResponseSchema>,
+      400
+    );
+  }
+
+  await stub.editIncident(incident);
 
   const user = await getUser(c.env, verifyInvitation.invitation.user);
   if (!user) {
@@ -90,18 +130,43 @@ export const handler: RouteHandler<Route, AppArgs> = async (c) => {
     );
   }
 
+  const currentIncident = await stub.getIncident(incident.id);
+  const result = mergeLWW({
+    local: currentIncident,
+    remote: incident,
+    ignore: INCIDENT_IGNORE,
+  });
+  if (!result.resolved) {
+    return c.json(
+      {
+        success: false,
+        code: "PatchIncidentEditInvalid",
+        error: {
+          name: "ValidationError",
+          message: "This incident could not be edited.",
+        },
+      } as const satisfies z.infer<typeof ErrorResponseSchema>,
+      409
+    );
+  }
+
+  await stub.editIncident(result.resolved);
+
   const sender: WebSocketSender = {
     id: user.key,
     name: user.name,
     type: "client",
   };
 
-  await stub.broadcast({ type: "remove_incident", id }, sender);
+  await stub.broadcast(
+    { type: "update_incident", incident: result.resolved },
+    sender
+  );
 
   return c.json(
     {
       success: true,
-      data: {},
+      data: result.resolved,
     } as const satisfies z.infer<typeof SuccessResponseSchema>,
     200
   );
