+++ worker/sync/src/api/sync/$sku/incident/put.ts
@@ -1,4 +1,8 @@
-import { ErrorResponseSchema, ErrorResponses, AppArgs } from "../../../router";
+import {
+  ErrorResponseSchema,
+  ErrorResponses,
+  AppArgs,
+} from "../../../../router";
 import { createRoute, RouteHandler } from "@hono/zod-openapi";
 import { z } from "zod/v4";
 import {
@@ -6,9 +10,9 @@ import {
   verifySignature,
   VerifySignatureHeadersSchema,
   verifyUser,
-} from "../../../utils/verify";
+} from "../../../../utils/verify";
 import { Incident, IncidentSchema, WebSocketSender } from "@referee-fyi/share";
-import { getUser } from "../../../utils/data";
+import { getUser } from "../../../../utils/data";
 export const ParamsSchema = z.object({
   sku: z.string(),
 });
