+++ worker/sync/src/api/sync/$sku/invite/delete.ts
@@ -1,4 +1,8 @@
-import { ErrorResponseSchema, ErrorResponses, AppArgs } from "../../../router";
+import {
+  ErrorResponseSchema,
+  ErrorResponses,
+  AppArgs,
+} from "../../../../router";
 import { createRoute, RouteHandler } from "@hono/zod-openapi";
 import { z } from "zod/v4";
 import {
@@ -6,7 +10,7 @@ import {
   verifySignature,
   VerifySignatureHeadersSchema,
   verifyUser,
-} from "../../../utils/verify";
+} from "../../../../utils/verify";
 import {
   Invitation,
   WebsocketServerUserRemoveMessage,
@@ -16,7 +20,7 @@ import {
   deleteInvitation,
   getInvitation,
   getUser,
-} from "../../../utils/data";
+} from "../../../../utils/data";
 export const ParamsSchema = z.object({
   sku: z.string(),
 });
