+++ worker/sync/src/api/sync/$sku/invite/put.ts
@@ -1,4 +1,8 @@
-import { ErrorResponseSchema, ErrorResponses, AppArgs } from "../../../router";
+import {
+  ErrorResponseSchema,
+  ErrorResponses,
+  AppArgs,
+} from "../../../../router";
 import { createRoute, RouteHandler } from "@hono/zod-openapi";
 import { z } from "zod/v4";
 import {
@@ -7,9 +11,9 @@ import {
   verifySignature,
   VerifySignatureHeadersSchema,
   verifyUser,
-} from "../../../utils/verify";
+} from "../../../../utils/verify";
 import { Invitation } from "@referee-fyi/share";
-import { getInvitation, setInvitation } from "../../../utils/data";
+import { getInvitation, setInvitation } from "../../../../utils/data";
 export const ParamsSchema = z.object({
   sku: z.string(),
 });
