+++ worker/sync/src/api/sync/$sku/request/put.ts
@@ -1,12 +1,16 @@
-import { ErrorResponseSchema, ErrorResponses, AppArgs } from "../../../router";
+import {
+  ErrorResponseSchema,
+  ErrorResponses,
+  AppArgs,
+} from "../../../../router";
 import { createRoute, RouteHandler } from "@hono/zod-openapi";
 import { z } from "zod/v4";
 import {
   verifySignature,
   VerifySignatureHeadersSchema,
   verifyUser,
-} from "../../../utils/verify";
-import { getInvitation, setRequestCode } from "../../../utils/data";
+} from "../../../../utils/verify";
+import { getInvitation, setRequestCode } from "../../../../utils/data";
 export const ParamsSchema = z.object({
   sku: z.string(),
 });
