+++ worker/sync/src/api/sync/user.ts
@@ -1,9 +1,12 @@
 import { createRoute, RouteHandler } from "@hono/zod-openapi";
-import { ErrorResponseSchema, ErrorResponses, AppArgs } from "../router";
+import { ErrorResponseSchema, ErrorResponses, AppArgs } from "../../router";
 import { z } from "zod/v4";
-import { verifySignature, VerifySignatureHeadersSchema } from "../utils/verify";
+import {
+  verifySignature,
+  VerifySignatureHeadersSchema,
+} from "../../utils/verify";
 import { UserSchema } from "@referee-fyi/share";
-import { getUser, isSystemKey, setUser } from "../utils/data";
+import { getUser, isSystemKey, setUser } from "../../utils/data";
 export const QuerySchema = z.object({
   name: z.string().min(1, "Name is required"),
 });
