+++ worker/sync/src/routes.ts
@@ -11,24 +11,24 @@ import * as api_integration_v1_$sku_incident_json from "./api/integration/v1/$sk
 import * as api_integration_v1_$sku_incident_pdf from "./api/integration/v1/$sku/incident/pdf";
 import * as api_integration_v1_$sku_incident_csv from "./api/integration/v1/$sku/incident/csv";
 import * as api_integration_v1_$sku_asset from "./api/integration/v1/$sku/asset";
-import * as api_meta_location from "./api/meta/location";
-import * as api_user from "./api/user";
-import * as api_$sku_invitation from "./api/$sku/invitation";
-import * as api_$sku_invite_put from "./api/$sku/invite/put";
-import * as api_$sku_invite_delete from "./api/$sku/invite/delete";
-import * as api_$sku_list from "./api/$sku/list";
-import * as api_$sku_request_put from "./api/$sku/request/put";
-import * as api_$sku_request_get from "./api/$sku/request/get";
-import * as api_$sku_create from "./api/$sku/create";
-import * as api_$sku_accept from "./api/$sku/accept";
-import * as api_$sku_asset_upload_url from "./api/$sku/asset/upload_url";
-import * as api_$sku_asset_url from "./api/$sku/asset/url";
-import * as api_$sku_asset_preview_url from "./api/$sku/asset/preview_url";
-import * as api_$sku_incident_put from "./api/$sku/incident/put";
-import * as api_$sku_incident_patch from "./api/$sku/incident/patch";
-import * as api_$sku_incident_delete from "./api/$sku/incident/delete";
-import * as api_$sku_data from "./api/$sku/data";
-import * as api_$sku_join from "./api/$sku/join";
+import * as api_meta_location from "./api/sync/meta/location";
+import * as api_user from "./api/sync/user";
+import * as api_$sku_invitation from "./api/sync/$sku/invitation";
+import * as api_$sku_invite_put from "./api/sync/$sku/invite/put";
+import * as api_$sku_invite_delete from "./api/sync/$sku/invite/delete";
+import * as api_$sku_list from "./api/sync/$sku/list";
+import * as api_$sku_request_put from "./api/sync/$sku/request/put";
+import * as api_$sku_request_get from "./api/sync/$sku/request/get";
+import * as api_$sku_create from "./api/sync/$sku/create";
+import * as api_$sku_accept from "./api/sync/$sku/accept";
+import * as api_$sku_asset_upload_url from "./api/sync/$sku/asset/upload_url";
+import * as api_$sku_asset_url from "./api/sync/$sku/asset/url";
+import * as api_$sku_asset_preview_url from "./api/sync/$sku/asset/preview_url";
+import * as api_$sku_incident_put from "./api/sync/$sku/incident/put";
+import * as api_$sku_incident_patch from "./api/sync/$sku/incident/patch";
+import * as api_$sku_incident_delete from "./api/sync/$sku/incident/delete";
+import * as api_$sku_data from "./api/sync/$sku/data";
+import * as api_$sku_join from "./api/sync/$sku/join";
 
 const app = new OpenAPIHono<AppArgs>();
